Nullable is now a property of a type, rather than a field.
Summary:
This diff is in preparation for changes to default & optional handling. In my cleanup, I took the opportunity to make optional a property of a type rather than a field. This results in few noticeable/practical changes in Babel since I'm explicitly disallowing option types in route signatures and variants. In the future, we can think about allowing option types in these places.

Test Plan: Modified tests.

Reviewers: kannan