Re-added support for Float32/Float64.
Summary:
I thought it might be worth it to have Float32 (Proto has it, and plenty of
languages we'll be supporting still differentiate between float/double).

A more restrictive min_value and max_value can be specified, which we/Carousel
could use for sanity checking longitude and latitude values for GPS coordinates.

Test Plan: Added tests

Reviewed By: guido