Stricter encoding and decoding.
Summary:
- A nullable union member is now encoded as a symbol when unspecified.
- "null" is rejected as an encoding of a nullable union member value that's
  unspecified.
- "null" is rejected as an encoding of a nullable struct field that's
  unspecified. Field must be absent.

Note that these changes have nothing to do with the "strict" keyword
for json_encode/json_decode.

If these constraints make serializing or deserializing difficult in Java or Swift,
please let me know and I'll make adjustments.

Also, notice that I'm enforcing that there's only one way to do things, which
is that "null" cannot appear in the serialized form for a struct or union. If
you think this is too strict, let me know.

Test Plan: Added and adjusted tests

Reviewed By: guido