Removed ForeignRef wrapper in favor of namespace attribute on CompositeType.
Summary:
Previously, using an imported type resulted in a wrapper: ForeignRef(DataType).
The wrapper has been removed. Generators are now expected to check the
DataType.namespace attribute to see what namespace the DataType was defined in,
and if foreign in the current context, take the necessary action.

Test Plan: Updated tests.

Reviewed By: kannan