JavaScript: Generate JSDoc for Stone data types
This commit adds a js_types generator that generates JSDoc @typedefs for Stone
struct and union types. JSDoc annotations for routes, which the js_client
generator generates, now reference these new @typedefs. As a consequence of this
change, I have migrated the --extra-args parameter from js_client to js_types,
since that parameter augments specific types with additional fields.

Below, I detail how the generator maps each Stone data type to JSDoc.

Namespaces
----------

To preserve namespaces, each user-defined type is prefixed with its Pascal-cased
namespace. For example, the Metadata struct in the files namespace is emitted as
FilesMetadata.

Basic Types
-----------

Most basic types are emitted as-before, except now using lowercase type names
(e.g., number instead of Number). This convention is consistent with JSDoc's
own documentation (see http://usejsdoc.org/tags-type.html), as well as
alternative JavaScript type systems (e.g., TypeScript).

A few other adjustments have been made, however:

* Void type is emitted as void rather than null.
  * JSDoc does not have an official void type, but ESLint's valid-jsdoc linting
    rule recognizes the void type. This is also consistent with TypeScript.
  * valid-jsdoc documentation: http://eslint.org/docs/rules/valid-jsdoc
* List types are emitted as Array.<ElementType> rather than just Array.
* Timestamps are emitted as Timestamp, which is JSDoc @typedef'd to string.
  * Previously, these were incorrectly emitted as "Object".

Aliases
-------

All aliases are unwrapped and emitted as their base type. If desired in the
future, every alias can be @typedef'd into their actual type, but that may
be confusing to developers perusing the documentation.

Structs
-------

Each struct is mapped to a JSDoc @typedef. Each field is emitted as a @property
on that @typedef. Unfortunately, JSDoc typedefs do not support inheritance like
JSDoc @classes, so every struct @typedef contains the fields from the struct's
parent type(s).

(Note: @class would be inappropriate to use for Stone data types, as these are
expected to be nominal types backed by a constructor.)

More information on typedefs here: http://usejsdoc.org/tags-typedef.html

Unions
------

Each union is mapped to a JSDoc @typedef with the appropriate fields. All
unions have a .tag property that contains a string value indicating the
union's variant. The JSDoc lists the type of .tag as the union of all known
string values (e.g., 'foo' | 'bar'). Strictly speaking, open unions may have
more possible values than listed; should this subtlety become an issue, the
generator could be modified to emit the generic string type in the union (e.g.,
'foo' | 'bar' | string).

All of the fields on the union that are only present on a particular variant
are marked as an optional @typedef @property, with a description mentioning
which .tag value it is tied to.

If a union uses inheritance, the @typedef simply incorporates all of the options
from the parent type(s).

Struct Polymorphism
-------------------

If a struct enumerates its subtypes, the generator emits an annotation for
the struct's .tag field that lets developers discriminate between the possible
subtypes. Like with unions, the .tag field's type is a union of all known
possible values (e.g., 'foo' | 'bar'). All subtypes contain an *optional* .tag
field typed with that subtype's tag value, since the .tag field may be omitted
when the subtype is referenced directly.

Any references to a struct that enumerates its subtypes are replaced with a
union of all possible subtypes. For example, consider the following Stone
types:

```
struct Resource
    union
        file File
        folder Folder

    path String

struct ResourceHolder
    value Resource
```

In the JSDoc, the `value` field on `ResourceHolder` is typed as `File | Folder |
 Resource`. If `Resource` were a closed union, then it would be typed as `File |
 Folder`.

Nullable Types
--------------

Nullable types are mapped to optional parameters, indicating that the
particular field may or may not be present on the data type.

Routes
------

js_client emits a method for each route that accepts a single parameter
corresponding to that route's Arg type and returns a Promise that resolves
to the route's Result type (or rejects with the route's Error type, wrapped
in a hardcoded Error object).

js_client now emits proper JSDoc annotations for each of these types. The
arg parameter references the appropriate type name emitted by js_types, and the
return value is a Promise parameterized with the result type and the error
type (`Promise.<ResultType, Error.<ErrorType>>`), where `Error` is the hardcoded
object that wraps an API error.