Fix publish action (#264)
The latest release didn't successfully publish; this fixes the errors from the GitHub actions.

For Python 2, encoding is not a valid keyword for open, so removed that. And for Python 3, PyPi returned a 400 status code saying Development Status :: 5 - Stable is an invalid input, so looked up the new string and added that. Also temporarily updated the workflow to run on PR, it successfully ran and published before this landed.

Co-authored-by: Scott Erickson <serickson@dropbox.com>