Add --attribute-comment argument to JS and TSD clients (#268)
Add --attribute-comment field to js_client.py and tsd_client.py, like the one added to python_client.py in #262. This is a prerequisite for dropbox/dropbox-sdk-js#891 which adds scopes to the outputted clients for easier reference.

Also adds appropriate testing and updates the docstring for the python_client.py, making it more succinct and aligned with the other two.