Version compatible inspect handling for class parameters
`getargspec` was deprecated in Python 3.0, but finally removed in Python
3.11. For reference: https://github.com/Pylons/pyramid/issues/3688

That seems like a valid amount of time.

We can do our best effort and if the user is running a Python version
that still has `getargspec` use that, but if they have `getfullargspec`
which is available in Python 3.11, we can use that.